pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distributed-training-framework"
version = "0.2.0"
description = "A small, educational distributed training framework built on PyTorch."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Dhanya Jayagopal" }]
dependencies = [
    "torch>=2.1",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
vision = ["torchvision>=0.16"]
dev = [
    "pytest>=7.4",
    "pytest-timeout>=2.2",
    "ruff>=0.5",
]

[project.scripts]
dtf-train = "trainer.cli:main"

[tool.setuptools.packages.find]
include = ["trainer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 180
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
